move nodeform into its own file
